Create doublyLinkList.cpp
Write a C program to support the following opearations on a doubly linked list where
each
node consists of integers.
a. Create a doubly linked list by adding each node at the front.
b. Insert a new node to the left of the node whose key value is read as an input
c. Delete the node of a given data,if it is found,otherwise display appropriate message.
d. Display the contents of the list.